Refactor constraint handling and add distance constraints.
Renamed constraint methods for clarity and simplified Jacobian computations. Added support for distance constraints and updated the `Solver` to handle constraint matrices. Implemented new methods in `DynamicSystem` to manage and integrate constraints effectively.

+++ include/core/types.h
@@ -4,10 +4,11 @@
 
 #include <cstdint>
 #include <Eigen/Dense>
+#include "utils/Logger.h"
 
 namespace Neutron {
 
-using UniqueID = std::uint32_t;
+using UniqueID = std::uint8_t;
 
 // Common vector/matrix types
 using Vector3d = Eigen::Vector3d;
